+++ cmd/import.go
@@ -16,60 +16,7 @@ import (
 	"github.com/konradit/mmt/pkg/utils"
 	"github.com/olekukonko/tablewriter"
 	"github.com/spf13/cobra"
-	"github.com/spf13/viper"
 )
-
-func get_flag_string(cmd *cobra.Command, name string) string {
-	value, err := cmd.Flags().GetString(name)
-        if err != nil {
-                cui.Error("Problem parsing "+name, err)
-        }
-        if value == "" {
-          value = viper.GetString("import."+name)
-        }
-        return value
-}
-
-func get_flag_slice(cmd *cobra.Command, name string) []string {
-	value, err := cmd.Flags().GetStringSlice(name)
-        if err != nil {
-                cui.Error("Problem parsing "+name, err)
-        }
-        if len(value) == 0 {
-          value = viper.GetStringSlice("import."+name)
-        }
-        return value
-}
-
-func get_flag_int(cmd *cobra.Command, name string, default_int string) int {
-	value, err := cmd.Flags().GetString(name)
-        if err != nil {
-                cui.Error("Problem parsing "+name, err)
-        }
-        if value == "" {
-          value = viper.GetString("import."+name)
-        }
-        if value == "" {
-          value = default_int
-        }
-        int1, err := strconv.Atoi(value)
-        return int1
-}
-
-func get_flag_bool(cmd *cobra.Command, name string, default_bool string) bool {
-	value, err := cmd.Flags().GetString(name)
-        if err != nil {
-                cui.Error("Problem parsing "+name, err)
-        }
-        if value == "" {
-          value = viper.GetString("import."+name)
-        }
-        if value == "" {
-          value = default_bool
-        }
-        bool1, err := strconv.ParseBool(value)
-        return bool1
-}
 
 var importCmd = &cobra.Command{
 	Use:   "import",
